fix: создание пакета constants для Enum'ов, написание тестов для проверки работы EventService, добавление проверки методов addLike и deleteLike

+++ src/main/java/ru/yandex/practicum/storage/db/FilmDbStorage.java
@@ -7,6 +7,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Component;
+import ru.yandex.practicum.exception.AlreadyExistException;
 import ru.yandex.practicum.exception.UnknownDataException;
 import ru.yandex.practicum.model.film.Director;
 import ru.yandex.practicum.model.film.Film;
@@ -199,6 +200,10 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public void addLike(int filmId, int userId) {
+        if (isLikeExists(filmId, userId)){
+            log.warn("У фильма с идентификатором {} уже стоит лайк от пользователя с идентификатором {}.", filmId, userId);
+            throw new AlreadyExistException("У фильма уже стоит лайк от данного пользователя.");
+        }
         jdbcTemplate.update(
                 "INSERT INTO FILM_LIKES VALUES ( ?,? )",
                 filmId,
@@ -208,6 +213,10 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public void deleteLike(int filmId, int userId) {
+        if (!isLikeExists(filmId, userId)){
+            log.warn("У фильма с идентификатором {} нет лайка от пользователя с идентификатором {}.", filmId, userId);
+            throw new AlreadyExistException("Невозможно удалит лайк от пользователя от которого нет лайка");
+        }
         jdbcTemplate.update(
                 "DELETE FROM FILM_LIKES WHERE FILM_ID= ? AND USER_ID = ?",
                 filmId,
@@ -284,7 +293,7 @@ public class FilmDbStorage implements FilmStorage {
     }
 
     @Override
-    public List<Film> getFilmsDirectorSortedByLikes(int directorId){
+    public List<Film> getFilmsDirectorSortedByLikes(int directorId) {
         String sqlQuery = "SELECT *, M.NAME AS mpa_name from film join MPA M on M.MPA_ID = FILM.MPA_ID " +
                 "LEFT JOIN film_likes fl ON film.film_id = fl.film_id " +
                 "WHERE film.film_id IN (" +
@@ -345,5 +354,12 @@ public class FilmDbStorage implements FilmStorage {
             log.warn("Фильм с идентификатором {} не найден.", id);
             throw new UnknownDataException("Фильм c id = " + id + " не найден");
         }
+    }
+
+    @Override
+    public boolean isLikeExists ( int filmId, int userId){
+        String checkQuery = "SELECT * FROM FILM_LIKES WHERE FILM_ID = ? AND USER_ID = ? ";
+        SqlRowSet filmRows = jdbcTemplate.queryForRowSet(checkQuery, filmId, userId);
+        return filmRows.next();
     }
 }
